refactor(gateway): move request client ip resolution to net
(cherry picked from commit 1d986f1c01508f6efdb25123d5b464ccb96f4068)
